--- src/main/java/tqs/sparkflow/stationservice/dto/BaseStationFields.java
@@ -1,129 +1,0 @@
-package tqs.sparkflow.stationservice.dto;
-
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-/**
- * Base class containing common fields for station-related DTOs.
- */
-public abstract class BaseStationFields<T> {
-    protected T id;
-    protected String name;
-    protected String address;
-    protected String city;
-    protected String country;
-    protected Double latitude;
-    protected Double longitude;
-    protected String connectorType;
-    protected Integer minPower;
-    protected Integer maxPower;
-    protected Boolean isOperational;
-    protected String status;
-    protected Double price;
-
-    public T getId() {
-        return id;
-    }
-
-    public void setId(T id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    public Double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(Double latitude) {
-        this.latitude = latitude;
-    }
-
-    public Double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(Double longitude) {
-        this.longitude = longitude;
-    }
-
-    public String getConnectorType() {
-        return connectorType;
-    }
-
-    public void setConnectorType(String connectorType) {
-        this.connectorType = connectorType;
-    }
-
-    public Integer getMinPower() {
-        return minPower;
-    }
-
-    public void setMinPower(Integer minPower) {
-        this.minPower = minPower;
-    }
-
-    public Integer getMaxPower() {
-        return maxPower;
-    }
-
-    public void setMaxPower(Integer maxPower) {
-        this.maxPower = maxPower;
-    }
-
-    public Boolean getIsOperational() {
-        return isOperational;
-    }
-
-    public void setIsOperational(Boolean isOperational) {
-        this.isOperational = isOperational;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public Double getPrice() {
-        return price;
-    }
-
-    public void setPrice(Double price) {
-        this.price = price;
-    }
-}
